Intersect a join list
For the select semantics, we need to remove the elements that are
not part of the domain. In the codegen we ask to sort both candidate
lists (lhs & rhs), then we intersect the two lists through the
GRAPHinterjoinlist. The function performs a sequential merge scan. If
both candidate lists are equal, then it returns its input but the rhs
unaltered. Otherwise it returns the intersection of the elements with
the proper elements in the edge src/dst filtered out.